+++ src/pages/index.ts
@@ -26,7 +26,7 @@ import PayHistory from "./PayHistory";
 import AdminAccount from "./admin/AdminAccount";
 import AdminCoin from "./admin/AdminCoin";
 import AdminDashboard from "./admin/AdminDashboard";
-import AdminFitnessList from "./admin/AdminFitnessList";
+import AdminFitnessList from "./admin/AdminFitnessList/AdminFitnessList";
 import AdminFitnessUpload from "./admin/AdminFitnessUpload";
 import AdminNotice from "./admin/AdminNotice";
 import AdminPayHistory from "./admin/AdminPayHistory";
